[ch-164493650] add MVC folder structure
